+++ backend/tracim_backend/models/context_models.py
@@ -43,7 +43,6 @@ from tracim_backend.models.event import EventTypeDatabaseParameters
 from tracim_backend.models.event import ReadStatus
 from tracim_backend.models.favorites import FavoriteContent
 from tracim_backend.models.roles import WorkspaceRoles
-from tracim_backend.models.todo import Todo
 
 
 class AboutModel(object):
@@ -1999,85 +1998,3 @@ class FavoriteContentInContext:
     @property
     def original_type(self) -> str:
         return self._favorite_content.original_type
-
-
-class TodoInContext(object):
-    """
-    Interface to get Todo data and related content
-
-    """
-
-    def __init__(self, todo: Todo, dbsession: Session, config: CFG, user: User,) -> None:
-        self.todo = todo
-        self.dbsession = dbsession
-        self.config = config
-        self._user = user
-
-    @property
-    def todo_id(self) -> int:
-        """
-        Todo id
-        :return: todo id as integer
-        """
-        return self.todo.todo_id
-
-    @property
-    def parent_id(self) -> int:
-        """
-        Todo parent id
-        :return: todo parent id as integer
-        """
-        return self.todo.content_id
-
-    @property
-    def assignee_id(self) -> int:
-        """
-        Todo assignee id
-        :return: todo assignee id as integer
-        """
-        return self.todo.assignee_id
-
-    @property
-    def status(self) -> str:
-        """
-        Todo status
-        :return: todo status as string
-        """
-        from tracim_backend.lib.core.content import ContentApi
-
-        content_api = ContentApi(
-            current_user=self._user, session=self.dbsession, config=self.config,
-        )
-        status = content_api.get_one(
-            self.todo.content_id, content_type=content_type_list.Todo.slug,
-        ).status
-
-        return status
-
-    @property
-    def raw_content(self) -> str:
-        """
-        Todo raw content
-        :return: todo raw content as string
-        """
-        from tracim_backend.lib.core.content import ContentApi
-
-        content_api = ContentApi(
-            current_user=self._user, session=self.dbsession, config=self.config,
-        )
-        raw_content = content_api.get_one(
-            self.todo.content_id, content_type=content_type_list.Todo.slug,
-        ).raw_content
-
-        return raw_content
-
-
-class TodoPath(object):
-    """
-    Paths parameters for todo: workspace id, content id, todo id
-    """
-
-    def __init__(self, workspace_id: int, content_id: int, todo_id: int) -> None:
-        self.content_id = content_id
-        self.workspace_id = workspace_id
-        self.todo_id = todo_id
